comment edu form, educator, edu on report, bubble chart, and bubble chart wrapper components
